Rename SRIOV_USERSPACE to vfio.
